+++ src/dependencies/parse-imports.test.ts
@@ -4,9 +4,12 @@ import { parse } from './../utils';
 describe('resolveImport', () => {
     it('should count require', () => {
         const result = parseImport(
-            parse(`
+            parse(
+                `
             const test1 = require('test-file1');
-        `),
+        `,
+                'filePath',
+            ),
         );
 
         expect(result.staticImports).toEqual([]);
@@ -15,9 +18,12 @@ describe('resolveImport', () => {
 
     it('should count es6 dynamic import', () => {
         const result = parseImport(
-            parse(`
+            parse(
+                `
             const test1 = import('test-file1');
-        `),
+        `,
+                'filePath',
+            ),
         );
 
         expect(result.staticImports).toEqual([]);
@@ -26,12 +32,15 @@ describe('resolveImport', () => {
 
     it('should count es6 imports', () => {
         const result = parseImport(
-            parse(`
+            parse(
+                `
             import { default as test1 } from 'test-file1';
             import { test2, test3 } from 'test-file2';
             import test4 from 'test-file3';
             import * as test5 from 'test-file4';
-        `),
+        `,
+                'filePath',
+            ),
         );
 
         expect(result.staticImports).toEqual([
@@ -45,11 +54,14 @@ describe('resolveImport', () => {
 
     it('should count es6 reexports', () => {
         const result = parseImport(
-            parse(`
+            parse(
+                `
             export { default as test1 } from 'test-file1';
             export { test2, test3 } from 'test-file2';
             export * from 'test-file3';
-        `),
+        `,
+                'filePath',
+            ),
         );
 
         expect(result.staticImports).toEqual(['test-file1', 'test-file2', 'test-file3']);
@@ -58,10 +70,13 @@ describe('resolveImport', () => {
 
     it('should not count exports', () => {
         const result = parseImport(
-            parse(`
+            parse(
+                `
             export default () => {};
             export const test = () => {};
-        `),
+        `,
+                'filePath',
+            ),
         );
 
         expect(result.staticImports).toEqual([]);
@@ -70,10 +85,13 @@ describe('resolveImport', () => {
 
     it('should ignore dynamic imports', () => {
         const result = parseImport(
-            parse(`
+            parse(
+                `
             const test1 = require(foo);
             const test2 = import(foo);
-        `),
+        `,
+                'filePath',
+            ),
         );
 
         expect(result.staticImports).toEqual([]);