+++ types/missing.d.ts
@@ -1,5 +1,4 @@
 declare module 'cosmiconfig'; // does not exist
-declare module '@babel/parser'; // does not exist
 declare module 'is-ci' {
     var isCi: boolean;
     export = isCi;